+++ actstream/migrations/0001_initial.py
@@ -2,13 +2,14 @@
 from __future__ import unicode_literals
 
 from django.db import models, migrations
+import django.db.models.deletion
 import django.utils.timezone
 from django.conf import settings
 
 from actstream.settings import USE_JSONFIELD
 
 if USE_JSONFIELD:
-    from jsonfield.fields import JSONField as DataField
+    from jsonfield_compat.fields import JSONField as DataField
 else:
     DataField = models.TextField
 
@@ -33,9 +34,9 @@ class Migration(migrations.Migration):
                 ('timestamp', models.DateTimeField(db_index=True, default=django.utils.timezone.now, help_text='')),
                 ('public', models.BooleanField(db_index=True, default=True, help_text='')),
                 ('data', DataField(blank=True, null=True, help_text='')),
-                ('action_object_content_type', models.ForeignKey(blank=True, null=True, help_text='', related_name='action_object', to='contenttypes.ContentType')),
-                ('actor_content_type', models.ForeignKey(help_text='', related_name='actor', to='contenttypes.ContentType')),
-                ('target_content_type', models.ForeignKey(blank=True, null=True, help_text='', related_name='target', to='contenttypes.ContentType')),
+                ('action_object_content_type', models.ForeignKey(on_delete=django.db.models.deletion.CASCADE, blank=True, null=True, help_text='', related_name='action_object', to='contenttypes.ContentType')),
+                ('actor_content_type', models.ForeignKey(on_delete=django.db.models.deletion.CASCADE, help_text='', related_name='actor', to='contenttypes.ContentType')),
+                ('target_content_type', models.ForeignKey(on_delete=django.db.models.deletion.CASCADE, blank=True, null=True, help_text='', related_name='target', to='contenttypes.ContentType')),
             ],
             options={
                 'ordering': ('-timestamp',),
@@ -48,8 +49,8 @@ class Migration(migrations.Migration):
                 ('object_id', models.CharField(max_length=255, db_index=True, help_text='')),
                 ('actor_only', models.BooleanField(verbose_name='Only follow actions where the object is the target.', default=True, help_text='')),
                 ('started', models.DateTimeField(db_index=True, default=django.utils.timezone.now, help_text='')),
-                ('content_type', models.ForeignKey(help_text='', to='contenttypes.ContentType')),
-                ('user', models.ForeignKey(help_text='', to=settings.AUTH_USER_MODEL)),
+                ('content_type', models.ForeignKey(on_delete=django.db.models.deletion.CASCADE, help_text='', to='contenttypes.ContentType')),
+                ('user', models.ForeignKey(on_delete=django.db.models.deletion.CASCADE, help_text='', to=settings.AUTH_USER_MODEL)),
             ],
         ),
         migrations.AlterUniqueTogether(