+++ actstream/views.py
@@ -1,5 +1,4 @@
-from django.shortcuts import render_to_response, get_object_or_404
-from django.template import RequestContext
+from django.shortcuts import get_object_or_404, render
 from django.http import HttpResponseRedirect, HttpResponse
 
 from django.contrib.auth.decorators import login_required
@@ -45,10 +44,16 @@ def stream(request):
     Index page for authenticated user's activity stream. (Eg: Your feed at
     github.com)
     """
-    return render_to_response('actstream/actor.html', {
-        'ctype': ContentType.objects.get_for_model(User),
-        'actor': request.user, 'action_list': models.user_stream(request.user)
-    }, context_instance=RequestContext(request))
+
+    return render(
+        request,
+        'actstream/actor.html',
+        {
+            'ctype': ContentType.objects.get_for_model(User),
+            'actor': request.user,
+            'action_list': models.user_stream(request.user)
+        }
+    )
 
 
 def followers(request, content_type_id, object_id):
@@ -58,39 +63,62 @@ def followers(request, content_type_id, object_id):
     """
     ctype = get_object_or_404(ContentType, pk=content_type_id)
     instance = get_object_or_404(ctype.model_class(), pk=object_id)
-    return render_to_response('actstream/followers.html', {
-        'followers': models.followers(instance), 'actor': instance
-    }, context_instance=RequestContext(request))
+
+    return render(
+        request,
+        'actstream/followers.html',
+        {
+            'followers': models.followers(instance),
+            'actor': instance
+        }
+    )
 
 
 def following(request, user_id):
     """
-    Returns a list of actors that the user identified by ``user_id`` is following (eg who im following).
+    Returns a list of actors that the user identified by ``user_id``
+    is following (eg who im following).
     """
     instance = get_object_or_404(User, pk=user_id)
-    return render_to_response('actstream/following.html', {
-        'following': models.following(instance), 'user': instance
-    }, context_instance=RequestContext(request))
+    return render(
+        request,
+        'actstream/following.html',
+        {
+            'following': models.following(instance),
+            'user': instance
+        }
+    )
 
 
 def user(request, username):
     """
     ``User`` focused activity stream. (Eg: Profile page twitter.com/justquick)
     """
-    instance = get_object_or_404(User, **{'is_active': True, compat.username_field(): username})
-    return render_to_response('actstream/actor.html', {
-        'ctype': ContentType.objects.get_for_model(User),
-        'actor': instance, 'action_list': models.user_stream(instance)
-    }, context_instance=RequestContext(request))
+    instance = get_object_or_404(
+        User,
+        **{'is_active': True, compat.username_field(): username}
+    )
+    return render(
+        request,
+        'actstream/actor.html',
+        {
+            'ctype': ContentType.objects.get_for_model(User),
+            'actor': instance, 'action_list': models.user_stream(instance)
+        }
+    )
 
 
 def detail(request, action_id):
     """
     ``Action`` detail view (pretty boring, mainly used for get_absolute_url)
     """
-    return render_to_response('actstream/detail.html', {
-        'action': get_object_or_404(models.Action, pk=action_id)
-    }, context_instance=RequestContext(request))
+    return render(
+        request,
+        'actstream/detail.html',
+        {
+            'action': get_object_or_404(models.Action, pk=action_id)
+        }
+    )
 
 
 def actor(request, content_type_id, object_id):
@@ -100,10 +128,15 @@ def actor(request, content_type_id, object_id):
     """
     ctype = get_object_or_404(ContentType, pk=content_type_id)
     instance = get_object_or_404(ctype.model_class(), pk=object_id)
-    return render_to_response('actstream/actor.html', {
-        'action_list': models.actor_stream(instance), 'actor': instance,
-        'ctype': ctype
-    }, context_instance=RequestContext(request))
+    return render(
+        request,
+        'actstream/actor.html',
+        {
+            'action_list': models.actor_stream(instance),
+            'actor': instance,
+            'ctype': ctype
+        }
+    )
 
 
 def model(request, content_type_id):
@@ -113,7 +146,12 @@ def model(request, content_type_id):
     """
     ctype = get_object_or_404(ContentType, pk=content_type_id)
     model_class = ctype.model_class()
-    return render_to_response('actstream/actor.html', {
-        'action_list': models.model_stream(model_class), 'ctype': ctype,
-        'actor': model_class
-    }, context_instance=RequestContext(request))
+    return render(
+        request,
+        'actstream/actor.html',
+        {
+            'action_list': models.model_stream(model_class),
+            'ctype': ctype,
+            'actor': model_class
+        }
+    )