+++ docs/source/changelog.rst
@@ -3,6 +3,18 @@
 Changelog
 =========
 
+0.6.4
+-----
+
+  - Dropped support for Django 1.4, 1.5 and 1.6
+  - Added explicit 'on_delete' arg to ForeignKey fields
+
+0.6.3
+-----
+
+  - MySQL monkey patch removal
+  - Scrutinizer CI integration
+
 0.6.2
 -----